Refactor: Simplify Hidden Single hint explanation
This commit refactors the "Hidden Single" hint explanation logic to be more concise and message-driven, removing complex cell-by-cell analysis.

The `HiddenSingleExplanation` class is moved to its own file (`sudoku-core/src/main/java/io/github/mgrablo/sudokucore/hints/strategies/HiddenSingleExplanation.kt`) and its implementation is significantly simplified. Instead of manually constructing sentences from various text parts and calculating affected cells, it now uses pre-defined, formatted strings from resources.

**Key Changes:**

*   **`HiddenSingleExplanation.kt`**:
    *   The class has been moved to its own dedicated file.
    *   The `generateStructuredHintExplanation` method is simplified to use `HintMessageFormatter` with new, specific string keys (`HIDDEN_SINGLE_STEP_1`, `_STEP_2`, `_STEP_3`), resulting in clearer, more direct hint steps.
    *   The previous logic for finding "affected cells" and constructing complex, multi-part `HintExplanationPart` lists has been removed.

*   **String Resources (`hint_strings.xml`, `HintStringKey.kt`)**:
    *   Added new string resources for the simplified Hidden Single explanation steps:
        *   `hint_hidden_single_step_1`: "Focus on the cell {0}!"
        *   `hint_hidden_single_step_2`: "In {0}, the only cell that can contain the number {1} is {2}."
        *   `hint_hidden_single_step_3`: "Therefore, the cell at {0} must contain {1}."
    *   Corresponding keys have been added to the `HintStringKey` enum.

*   **`AndroidHintStringProvider.kt`**:
    *   Updated to map the new `HIDDEN_SINGLE_STEP` keys to their respective Android string resources.

*   **`HintExplanationStrategy.kt`**:
    *   The old implementation of `HiddenSingleExplanation` has been removed from this file.
    *   Helper functions `getScopePartStringLocalized` and `getScopeTypeString` are now marked as `internal`.
